refactor: move app logic to app.ts
